Rename callbacks/base.py to callbacks/callback.py (#13031)
