+++ pytorch_lightning/callbacks/early_stopping.py
@@ -25,7 +25,7 @@ import numpy as np
 import torch
 
 import pytorch_lightning as pl
-from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.callbacks.callback import Callback
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.rank_zero import rank_zero_warn
