+++ pytorch_lightning/callbacks/gradient_accumulation_scheduler.py
@@ -23,7 +23,7 @@ Trainer also calls ``optimizer.step()`` for the last indivisible step number.
 from typing import Any, Dict
 
 import pytorch_lightning as pl
-from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.callbacks.callback import Callback
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
