+++ pytorch_lightning/callbacks/lambda_function.py
@@ -21,7 +21,7 @@ Create a simple callback on the fly using lambda functions.
 
 from typing import Callable, Optional
 
-from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.callbacks.callback import Callback
 
 
 class LambdaCallback(Callback):
@@ -29,7 +29,7 @@ class LambdaCallback(Callback):
     Create a simple callback on the fly using lambda functions.
 
     Args:
-        **kwargs: hooks supported by :class:`~pytorch_lightning.callbacks.base.Callback`
+        **kwargs: hooks supported by :class:`~pytorch_lightning.callbacks.callback.Callback`
 
     Example::
