+++ pytorch_lightning/callbacks/lr_monitor.py
@@ -26,7 +26,7 @@ from typing import Any, DefaultDict, Dict, List, Optional, Set, Tuple, Type
 from torch.optim.optimizer import Optimizer
 
 import pytorch_lightning as pl
-from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.callbacks.callback import Callback
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.rank_zero import rank_zero_deprecation, rank_zero_warn
 from pytorch_lightning.utilities.types import LRSchedulerConfig
