+++ pytorch_lightning/callbacks/model_summary.py
@@ -25,7 +25,7 @@ import logging
 from typing import List, Tuple
 
 import pytorch_lightning as pl
-from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.callbacks.callback import Callback
 from pytorch_lightning.utilities.model_summary import _format_summary_table, summarize
 
 log = logging.getLogger(__name__)
