+++ pytorch_lightning/callbacks/prediction_writer.py
@@ -20,7 +20,7 @@ Aids in saving predictions
 from typing import Any, Optional, Sequence
 
 import pytorch_lightning as pl
-from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.callbacks.callback import Callback
 from pytorch_lightning.utilities import LightningEnum
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
