+++ pytorch_lightning/callbacks/pruning.py
@@ -27,7 +27,7 @@ from torch import nn
 from typing_extensions import TypedDict
 
 import pytorch_lightning as pl
-from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.callbacks.callback import Callback
 from pytorch_lightning.core.module import LightningModule
 from pytorch_lightning.utilities.apply_func import apply_to_collection
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
