+++ pytorch_lightning/callbacks/quantization.py
@@ -25,7 +25,7 @@ from torch import Tensor
 from torch.quantization import FakeQuantizeBase
 
 import pytorch_lightning as pl
-from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.callbacks.callback import Callback
 from pytorch_lightning.utilities import _TORCH_GREATER_EQUAL_1_10, _TORCH_GREATER_EQUAL_1_11
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
