+++ pytorch_lightning/callbacks/stochastic_weight_avg.py
@@ -23,7 +23,7 @@ from torch import nn
 from torch.optim.swa_utils import SWALR
 
 import pytorch_lightning as pl
-from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.callbacks.callback import Callback
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.rank_zero import rank_zero_info, rank_zero_warn
 from pytorch_lightning.utilities.types import LRSchedulerConfig
