+++ pytorch_lightning/callbacks/timer.py
@@ -21,7 +21,7 @@ from datetime import timedelta
 from typing import Any, Dict, Optional, Union
 
 import pytorch_lightning as pl
-from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.callbacks.callback import Callback
 from pytorch_lightning.trainer.states import RunningStage
 from pytorch_lightning.utilities import LightningEnum
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
